requirements.txt
-e git+https://github.com/davechogan/AgileStories.git@96413915972a953fc75ff8aebdec39b9263ff641#egg=agilestories
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
boto3==1.35.97
botocore==1.35.97
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.6
frozenlist==1.5.0
h11==0.14.0
idna==3.10
jmespath==1.0.1
multidict==6.1.0
openai==0.28.1
propcache==0.2.1
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.20
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
